changing quote form on userprofile as well as finishing contact page
added button to hide/show quote forms as well as added a normal quote form for the business standard. added disabling of button through angular if the required fields are not met.
filled contact page out with community links as well as a form to message  owner with email form